Hiêu chỉnh dữ liệu trường thời gian trên chương trình, thời gian thực hiện hệ thống tự tính

+++ EVN.Api/Model/Request/XacNhanTroNgaiRequest.cs
@@ -98,8 +98,8 @@ namespace EVN.Api.Model.Request
         public string SO_NGAY_TH_ND { get; set; }
         public int TRANGTHAI_GQ { get; set; }
         public int TONG_CONGSUAT_CD { get; set; }
-        public int DGCD_TH_CHUONGTRINH { get; set; }
-        public int DGCD_TH_DANGKY { get; set; }
+        public double DGCD_TH_CHUONGTRINH { get; set; }
+        public double DGCD_TH_DANGKY { get; set; }
         public int DGCD_KH_PHANHOI { get; set; }
         public int DGYC_DK_DEDANG { get; set; }
         public int DGYC_XACNHAN_NCHONG_KTHOI { get; set; }
